+++ src/code42cli/cmds/alert_rules.py
@@ -15,6 +15,9 @@ from code42cli.file_readers import read_csv_arg
 from code42cli.options import format_option
 from code42cli.options import sdk_options
 from code42cli.output_formats import OutputFormatter
+from code42cli.util import deprecation_warning
+
+DEPRECATION_TEXT = "Incydr functionality is deprecated. Use the Incydr CLI instead."
 
 
 class AlertRuleTypes:
@@ -35,7 +38,8 @@ _HEADER_KEYS_MAP["isEnabled"] = "Enabled"
 @click.group(cls=OrderedGroup)
 @sdk_options(hidden=True)
 def alert_rules(state):
-    """Manage users associated with alert rules."""
+    """DEPRECATED - Manage users associated with alert rules."""
+    deprecation_warning(DEPRECATION_TEXT)
     pass